+++ src/App.atest.tsx
@@ -1,0 +1,41 @@
+// import { render, screen } from '@testing-library/react';
+// import { BrowserRouter, MemoryRouter } from 'react-router-dom';
+// import userEvent from '@testing-library/user-event';
+// import React from 'react';
+// import App from './App';
+
+describe('App', () => {
+  // it('full App rendering', () => {
+  //   render(<App />, { wrapper: BrowserRouter });
+  //   expect(screen.getByText(/Cocktail Search/i)).toBeInTheDocument();
+  // });
+  // it('home navigation', async () => {
+  //   render(<App />, { wrapper: BrowserRouter });
+  //   const user = userEvent.setup();
+  //   await user.click(screen.getByText(/main page/i));
+  //   expect(screen.getByText(/Cocktail Search/i)).toBeInTheDocument();
+  // });
+  // it('about us navigation', async () => {
+  //   render(<App />, { wrapper: BrowserRouter });
+  //   const user = userEvent.setup();
+  //   await user.click(screen.getByText(/about/i));
+  //   expect(screen.getByText(/About Page/i)).toBeInTheDocument();
+  // });
+  // it('form navigation', async () => {
+  //   render(<App />, { wrapper: BrowserRouter });
+  //   const user = userEvent.setup();
+  //   await user.click(screen.getByText(/create your cocktail/i));
+  //   expect(
+  //     screen.getByText(/Add your own cocktail recipe/i)
+  //   ).toBeInTheDocument();
+  // });
+  // test('landing on a bad page', () => {
+  //   const badRoute = '/some/bad/route';
+  //   render(
+  //     <MemoryRouter initialEntries={[badRoute]}>
+  //       <App />
+  //     </MemoryRouter>
+  //   );
+  //   expect(screen.getByText(/Not Found/i)).toBeInTheDocument();
+  // });
+});
